% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defExport.R
\name{defExport}
\alias{defExport}
\title{Create an export definition}
\usage{
defExport(login, type, fields = NULL, bulkFilter = list(),
  exportName = paste("Bulk API Export ", Sys.Date(), sep = ""))
}
\arguments{
\item{login}{An eloquaLogin object}

\item{type}{The type of export to create. Accepted values are contained in eloquaActivities}

\item{fields}{A vector of the desired field names. If null, defaults to all available fields}

\item{bulkFilter}{A named list containing filter parameters; example: list("ActivityDate" = c(operator = ">", value = "2015-11-20 00:00:00"))}

\item{exportName}{A name to give to the export. If null, defaults to "Bulk API Export + (Timestamp)"}
}
\value{
A list; return of POST request to the export definition API endpoint
}
\description{
Create an export definition
}

