% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloquaActivities.R
\docType{data}
\name{eloquaActivities}
\alias{eloquaActivities}
\title{Named list of activity objects available for export via Bulk API}
\format{An object of class \code{list} of length 10.}
\usage{
eloquaActivities
}
\description{
Contains the available Eloqua activity types and their respective fields
}
\keyword{datasets}
