% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractExportData.R
\name{extractExportData}
\alias{extractExportData}
\title{Parses through a list of lists from getExportData and returns data frame}
\usage{
extractExportData(export = list())
}
\arguments{
\item{export}{A list of lists from getExportData}
}
\value{
A dataframe
}
\description{
Parses through a list of lists from getExportData and returns data frame
}

