% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getExportData.R
\name{getExportData}
\alias{getExportData}
\title{Loop through the synced export data and return to a list of lists}
\usage{
getExportData(login, exportDefinition)
}
\arguments{
\item{login}{An eloquaLogin object}

\item{exportDefinition}{An httr POST return list created with defExport}
}
\value{
A list of lists containint the httr GET responses from the /activities/export/{id}/data endpoint
}
\description{
Loop through the synced export data and return to a list of lists
}

