% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSyncStatus.R
\name{getSyncStatus}
\alias{getSyncStatus}
\title{Get the status of a previously created sync}
\usage{
getSyncStatus(login, syncResponse)
}
\arguments{
\item{login}{An eloquaLogin object}

\item{syncResponse}{An httr POST return list created with startSyncExport}
}
\value{
An httr POST list with the status of the sync
}
\description{
Get the status of a previously created sync
}

