% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/startSyncExport.R
\name{startSyncExport}
\alias{startSyncExport}
\title{Begins syncing data for an export definition}
\usage{
startSyncExport(login, exportDefinition)
}
\arguments{
\item{login}{An eloquaLogin object}

\item{exportDefinition}{An httr POST return list created with defExport}
}
\value{
An httr POST list with the status of call and sync URI
}
\description{
Begins syncing data for an export definition
}

