% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syncExport.R
\name{syncExport}
\alias{syncExport}
\title{Complete cycle of creating a sync and checking until finished}
\usage{
syncExport(login, exportDefinition, maxTimeout = 180)
}
\arguments{
\item{login}{An eloquaLogin object}

\item{exportDefinition}{An httr POST return list created with defExport}

\item{maxTimeout}{Number of seconds to retry before timeout occurs}
}
\value{
A list of two lists, results from startSyncExport and getSyncStatus, after the sync is completed (or timeout occurs)
}
\description{
Complete cycle of creating a sync and checking until finished
}

