% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsgetCurrStudyByType.R
\name{wsgetCurrStudyByType}
\alias{wsgetCurrStudyByType}
\title{Loop through the synced export data and return to a list of lists}
\usage{
wsgetCurrStudyByType(login, currEloquaType, currBulkFilter)
}
\arguments{
\item{login}{An eloquaLogin object}

\item{currEloquaType}{Type Include - PageView,WebVisit,FormSubmit,FormSubmitNoRaw,EmailSend,EmailOpen,EmailClickthrough,Subscribe,Unsubscribe,Bounceback}
}
\description{
Loop through the synced export data and return to a list of lists
}
