% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eloquaLogin.R
\docType{class}
\name{eloquaLogin-class}
\alias{eloquaLogin}
\alias{eloquaLogin-class}
\title{Create a set of authentication credentials required for Eloqua}
\arguments{
\item{pod}{The pod your Eloqua instance resides in (used to create the UIR string)}

\item{company}{The company/name of your Eloqua instance}

\item{username}{Your username}

\item{password}{Your password}
}
\value{
An object of class eloquaLogin
}
\description{
Create a set of authentication credentials required for Eloqua
}

